Fix info macro test
